run.query bad param
I did this wrong. The way I did it before makes always set the conn to
reset.conn(). But sometimes I want to specify a different database, or
maybe I really can just get away with last_connection(). Doing it this
way, it soft defaults to reset.conn(), but gives me the ability to
override.