Factors fuck it up
Problem occurs when comparing data frames with only one other column. If
that other column is a factor, you're potentially comparing two
different sets with two different sets of levels. And so, you get an
error. I think this'll cast the whole vector as a character, and then
compare the character values.