Adding clip.tail
Adding function to trim a continuous series of values from the end of a
vector and replace them with nulls. Good for cleaning a vector that,
say, represents a marginal distribution which you want to turn
cumulative, but it has a bunch of zeros at the end that makes the cum
curve look flat.